fix remaining bench-harness lint failures (rust fmt, dart, swift, csharp)
The benchmark harness files were not passing each port's strict linter:

- rust: cargo fmt --check wanted examples/bench.rs reflowed
  (env-var chain and the path literal); applied cargo fmt.
- dart: bench/bench.dart had an unnecessary 'as int' cast on
  stringify().length (already int); removed it.
- swift: Sources/bench/main.swift line 69 exceeded SwiftLint's
  LineLength error threshold (220 chars); split the final print
  into short composed segments.
- csharp: bench/Program.cs tripped 27 analyzer rules under
  -warnaserror (IDE0040/0011/0022/0300/0039/0062/0005, CA1805/
  1852/1305). Add accessibility modifiers, braces and block
  bodies, seal the class, drop the redundant sink initialiser
  and unused usings, make the walk callback a static local
  function, and pass InvariantCulture to StringBuilder.Append.

All four benches still build and emit identical output (sink
values unchanged). Verified locally: cargo fmt --check, dart
analyze, dotnet build -warnaserror, and the swift/csharp bench
runs.

+++ dart/bench/bench.dart
@@ -64,7 +64,7 @@ void main() {
     ['clone', nodes, () => sink += s.clone(tree) != null ? 1 : 0],
     ['walk', nodes, () => s.walk(tree, before: cb)],
     ['merge', nodes, () => sink += s.merge([treeA, treeB]) != null ? 1 : 0],
-    ['stringify', nodes, () => sink += s.stringify(tree).length as int],
+    ['stringify', nodes, () => sink += s.stringify(tree).length],
     ['getpath', GP, () {
       var acc = 0;
       for (var i = 0; i < GP; i++) {
